Extract presentational views for SubmittedCodeTable and StudyRow
Each splits into a pure View (no data or session hooks, in its own file) plus a thin container that keeps the data plumbing: SubmittedCodeTableView under SubmittedCodeTable (useQuery + preview modal), and StudyRowView under StudyRow (useStudyStatus + the Clerk-reading StudyActionLink, now injected as a prop). Behavior is unchanged and existing tests pass; the views can now render in isolation in Ladle.
