+++ src/components/dashboard/studies-table/study-row.tsx
@@ -1,11 +1,8 @@
-import { DisplayStudyStatus } from '@/components/study/display-study-status'
-import { InfoTooltip } from '@/components/tooltip'
 import { useStudyStatus } from '@/hooks/use-study-status'
-import { TableTd, TableTr, Text, useMantineTheme } from '@mantine/core'
-import dayjs from 'dayjs'
 import { StudyActionLink } from './study-action-link'
 import { studyHasJobStatus } from '@/lib/studies'
 import { Audience, Scope, StudyRow as StudyRowType } from './types'
+import { StudyRowView } from './study-row-view'
 
 type StudyRowProps = {
     study: StudyRowType
@@ -22,7 +19,6 @@ function shouldHighlight(study: StudyRowType, audience: Audience): boolean {
 }
 
 export function StudyRow({ study, audience, scope, orgSlug }: StudyRowProps) {
-    const theme = useMantineTheme()
     const status = useStudyStatus({
         studyStatus: study.status,
         audience,
@@ -30,50 +26,15 @@ export function StudyRow({ study, audience, scope, orgSlug }: StudyRowProps) {
     })
 
     const isHighlighted = shouldHighlight(study, audience)
-    const highlightStyle = isHighlighted
-        ? { backgroundColor: `${theme.colors.purple[0]}80`, fontWeight: 600 }
-        : undefined
-
-    // Get the "Submitted To" value (researcher only)
-    const submittedTo = study.reviewingEnclaveName || study.orgName || ''
-
-    // Get the "Submitted By" value (reviewer only)
-    const submittedBy = study.createdBy
-
-    const reviewedByOrOrg = scope === 'org' ? (study.reviewerName ?? '-') : study.orgName
 
     return (
-        <TableTr fz={14} style={highlightStyle}>
-            {/* Study Name - common to all */}
-            <TableTd>
-                <InfoTooltip label={study.title} multiline maw={400}>
-                    <Text
-                        lineClamp={2}
-                        style={{ cursor: 'pointer', overflowWrap: 'break-word' }}
-                        size="sm"
-                        fw={isHighlighted ? 600 : undefined}
-                    >
-                        {study.title}
-                    </Text>
-                </InfoTooltip>
-            </TableTd>
-
-            {/* Last updated - common to all */}
-            <TableTd>{dayjs(study.lastUpdatedAt).format('MMM DD, YYYY')}</TableTd>
-
-            {/* Third column differs by audience */}
-            {audience === 'researcher' ? <TableTd>{submittedTo}</TableTd> : <TableTd>{submittedBy}</TableTd>}
-
-            {/* Fourth column - reviewer only has this extra column */}
-            {audience === 'reviewer' && <TableTd>{reviewedByOrOrg}</TableTd>}
-
-            {/* Status - common to all */}
-            <TableTd>
-                <DisplayStudyStatus status={status} />
-            </TableTd>
-
-            {/* Action Link - common to all */}
-            <TableTd ta="center">
+        <StudyRowView
+            study={study}
+            audience={audience}
+            scope={scope}
+            status={status}
+            isHighlighted={isHighlighted}
+            actionLink={
                 <StudyActionLink
                     study={study}
                     audience={audience}
@@ -81,7 +42,7 @@ export function StudyRow({ study, audience, scope, orgSlug }: StudyRowProps) {
                     orgSlug={orgSlug}
                     isHighlighted={isHighlighted}
                 />
-            </TableTd>
-        </TableTr>
+            }
+        />
     )
 }
